// server/src/services/extraction/entities/project_charter_details/types.ts
export interface ProjectCharterDetails {
    project_charter_id?: string;
    vision?: string;
    mission?: string;
    project_manager?: string;
    sponsor?: string;
    authority_level?: string;
    major_milestones?: string[];
    high_level_risks?: string[];
    critical_success_factors?: string[];
    source_document?: string;
    source_document_id?: string;
}